air-quality-workflow/requirements.txt
azure-functions
pyodbc
python-dotenv
azure-identity
